+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Added: class BE\::IndentedOut
 - Fixed: AST\::Interface\::get_all_bases ()
+- Fixed: pseudo interfaces may inherit other pseudo interfaces
 
 ### 1.0.1 28.03.2021